requirements.txt
astroid==2.3.3
autopep8==1.5.1
colorama==0.4.3
Django==2.0.6
django-cors-headers==2.2.0
django-crispy-forms==1.9.0
django-tinymce==3.0.2
djangorestframework==3.8.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.5.0
pylint==2.4.4
pytz==2018.4
six==1.14.0
wrapt==1.11.2
